api_gateway_service/requirements.txt
aiokafka==0.8.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
asyncpgsa==0.27.1
attrs==23.1.0
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
